Add isSameSession to Session.java and move it under session package
